requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Computer Vision
opencv-python>=4.8.0
albumentations>=1.3.0
pillow>=10.0.0

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Configuration & Logging
pyyaml>=6.0
tqdm>=4.65.0
wandb>=0.15.0  # Optional: for experiment tracking

# Metrics
torchmetrics>=1.0.0

# Utilities
einops>=0.6.0
timm>=0.9.0  # For pretrained backbones

# Development
jupyter>=1.0.0
ipython>=8.12.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0